refactor: winning lotto가 lotto를 상속해서 관련 로직을 lotto로 옮겨주고 생성자에서 lotto 자체를 주입받도록 만듦
