+++ src/main/java/lotto/domain/Lotto.java
@@ -2,16 +2,37 @@ package lotto.domain;
 
 import com.google.common.primitives.Ints;
 
-import java.util.List;
+import java.util.*;
 
 public class Lotto {
     public static final int LOTTO_START_INDEX = 0;
     public static final int LOTTO_SIZE = 6;
+    public static final String LOTTO_SIZE_ERROR = String.format("번호는 총 %d개 이어야 합니다.", LOTTO_SIZE);
+    public static final String DUPLICATE_ERROR = "중복되는 번호가 있습니다.";
 
-    private final List<LottoNumber> lottoNumbers;
+    protected final List<LottoNumber> lottoNumbers;
 
     public Lotto(List<LottoNumber> lottoNumbers) {
+        validateLottoSize(lottoNumbers);
+        validateDuplicates(lottoNumbers);
         this.lottoNumbers = lottoNumbers;
+    }
+
+    private void validateLottoSize(List<LottoNumber> numbers) {
+        if (numbers.size() != LOTTO_SIZE) {
+            throw new IllegalArgumentException(LOTTO_SIZE_ERROR);
+        }
+    }
+
+    private void validateDuplicates(List<LottoNumber> numbers) {
+        Set<LottoNumber> numberGroup = new HashSet<>(numbers);
+        if (numberGroup.size() != LOTTO_SIZE) {
+            throw new IllegalArgumentException(DUPLICATE_ERROR);
+        }
+    }
+
+    public List<LottoNumber> toList() {
+        return Collections.unmodifiableList(lottoNumbers);
     }
 
     public int countMatchingNumbers(List<LottoNumber> numbers) {
